Updates templates url to hit the templateCache
Adding the plugin to an angular application without including the template files at the specified path will prevent the plugin from working at all. The templateCache has probably been put there so the users of the plugin don't necessarily have to include the template files at that exact path.

However, the requested path and the cached path differ, so AngularJS makes a request to the template path and fails. This change updates the request template path to match the cache.